requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
tldextract==5.1.1
SQLAlchemy==2.0.27 